refactor: modularize Target
