+++ src/types.rs
@@ -1,35 +1,14 @@
 pub mod error;
 pub mod family;
+pub mod target;
 
 use clap::{Args, Parser, Subcommand, ValueEnum};
 use error::{Error, InvalidChip};
-use std::{fmt::Display, str::FromStr};
-
-#[derive(Debug, Clone, ValueEnum)]
-#[value()]
-pub enum Target {
-    Thumbv6,
-    Thumbv7,
-    Thumbv7e,
-    Thumbv7f,
-    Thumbv8,
-}
-
-impl Display for Target {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(match self {
-            Self::Thumbv6 => "thumbv6m-none-eabi",
-            Self::Thumbv7 => "thumbv7m-none-eabi",
-            Self::Thumbv7e => "thumbv7em-none-eabi",
-            Self::Thumbv7f => "thumbv7em-none-eabihf",
-            Self::Thumbv8 => "thumbv8m.main-none-eabihf",
-        })
-    }
-}
+use std::str::FromStr;
 
 pub(crate) struct Chip {
     pub family: family::Family,
-    pub target: Target,
+    pub target: target::Target,
     pub name: String,
 }
 
@@ -39,7 +18,7 @@ impl FromStr for Chip {
     fn from_str(chip: &str) -> Result<Self, Self::Err> {
         use family::mem_region::MemRegion;
         use family::Family::*;
-        use Target::*;
+        use target::Target::*;
 
         let chips = [
             // nRF
